refactor: extract contentscript into modules
