+++ src/pages/content/modules/index.ts
@@ -1,0 +1,4 @@
+export * from './autoDetection';
+export * from './scriptInjection';
+export * from './messageHandler';
+export * from './eventHandler';
